Change file name and class name to match and be more appropriate!
